App restructured to use context in redux-like manner. Update item now functional
