+++ screens/EditTaskScreen.js
@@ -1,12 +1,14 @@
-import React, { useState } from 'react';
+import React, { useState, useContext } from 'react';
 import { TextInput, StyleSheet, Button, View } from 'react-native';
+import ItemsContext from './../contexts/ItemsContext';
 
 const EditTaskScreen = ({ route, navigation }) => {
   const { item } = route.params;
   const [editedtext, setEditedText] = useState(`${item.text}`);
+  values = useContext(ItemsContext);
 
   const pressHandler = () => {
-    navigation.navigate('To Do List', { text: editedtext });
+    navigation.navigate('To Do List', values.updateItem(item.id, editedtext));
   };
 
   navigation.setOptions({
